healthd: Reinitialize mChargerNames for every battery update
Booting up the device without usb, the kernel sets the usb power supply
type as UNKNOWN. The type of usb power supply changes at run-time as
various chargers are plugged in/out. However, healthd initilizes the
charger list only at bootup. Change it such that it checks for charger
type changes with every battery or usb uevent.

While at it, the kernel may have a power supply type which is not known
to healthd. This is perfectly fine. Update healthd to not print a
warning.

Change-Id: I2ec9f9a420ca61814d43c316b418ce94de3691bc

Former-commit-id: d50fc946399cbf763d6666a5855cd080d5d10c03
Former-commit-id: 7d0524fab1717bfaa880f0a24e600baee0edfe35
Former-commit-id: 1a43bb9821df5635805858be75b8f2b94ef0d9fd